Generic tests for talk filters extracted to seperate file
